+++ src/components/recurring-donations/1rupee/registration-info.tsx
