--- src/config/donation.ts
@@ -1,19 +1,0 @@
-import { STATIC_IMAGE_HOST } from "@/config/config";
-
-export const DonationDetails = {
-	accountName: "Aid Global Foundation",
-
-	accountNumber: "925020029188000",
-
-	upiID: "aidglobalfounda973306.rzp@icici",
-
-	bankName: "Axis Bank Ltd",
-
-	branchAddress: "Anjur phata bhiwandi Branch",
-
-	codeIFSC: "UTIB0003387",
-
-	codeMICR: "400211288",
-
-	qrSrc: `${STATIC_IMAGE_HOST}donation/qr.webp`,
-};
